+++ src/main/java/tornadofx/Builders.kt
@@ -1,5 +1,6 @@
 package tornadofx
 
+import javafx.beans.property.ObjectProperty
 import javafx.beans.property.Property
 import javafx.beans.value.ObservableValue
 import javafx.collections.ObservableList
@@ -11,6 +12,9 @@ import javafx.scene.text.Text
 import javafx.util.Callback
 import javafx.util.StringConverter
 import java.time.LocalDate
+import kotlin.reflect.KFunction
+import kotlin.reflect.KMutableProperty1
+import kotlin.reflect.KProperty1
 
 fun Pane.titledPane(title: String, node: Node): TitledPane {
     val pane = TitledPane(title, node)
@@ -39,6 +43,7 @@ fun GridPane.row(title: String? = null, op: Pane.() -> Unit) {
     // Create a new row in the GridPane and add the children added to the fake pane
     addRow(userData as Int, *fake.children.toTypedArray())
 }
+
 
 fun Pane.text(initialValue: String? = null, op: (Text.() -> Unit)? = null) = opcr(this, Text().apply { if (initialValue != null) text = initialValue }, op)
 fun Pane.text(property: Property<String>, op: (Text.() -> Unit)? = null) = text(op = op).apply {
@@ -127,6 +132,28 @@ fun Pane.vbox(spacing: Double? = null, children: Iterable<Node>? = null, op: (VB
 
 fun Pane.stackpane(initialChildren: Iterable<Node>? = null, op: (StackPane.() -> Unit)? = null) = opcr(this, StackPane().apply { if (initialChildren != null) children.addAll(initialChildren) }, op)
 fun Pane.gridpane(op: (GridPane.() -> Unit)? = null) = opcr(this, GridPane(), op)
+fun Pane.borderpane(op: (BorderPane.() -> Unit)? = null) = opcr(this,BorderPane(),op)
+
+fun BorderPane.top(op: Pane.() -> Unit) {
+    val vbox = VBox()
+    op(vbox)
+    top = if (vbox.children.size == 1) vbox.children[0] else vbox
+}
+fun BorderPane.bottom(op: Pane.() -> Unit) {
+    val vbox = VBox()
+    op(vbox)
+    bottom = if (vbox.children.size == 1) vbox.children[0] else vbox
+}
+fun BorderPane.left(op: Pane.() -> Unit) {
+    val vbox = VBox()
+    op(vbox)
+    left = if (vbox.children.size == 1) vbox.children[0] else vbox
+}
+fun BorderPane.right(op: Pane.() -> Unit) {
+    val vbox = VBox()
+    op(vbox)
+    right = if (vbox.children.size == 1) vbox.children[0] else vbox
+}
 
 /**
  * Add the given node to the pane, invoke the node operation and return the node
@@ -145,9 +172,45 @@ inline fun <S> Pane.tableview(op: (FXTableView<S>.() -> Unit)): FXTableView<S> {
 }
 
 class FXTableView<S> : TableView<S>() {
+    /**
+     * Create a column with a value factory that extracts the value from the given callback.
+     */
     fun <T> column(title: String, valueProvider: (TableColumn.CellDataFeatures<S, T>) -> ObservableValue<T>): TableColumn<S, T> {
         val column = TableColumn<S, T>(title)
         column.cellValueFactory = Callback { valueProvider(it) }
+        columns.add(column)
+        return column
+    }
+
+    /**
+     * Create a column with a value factory that extracts the value from the given mutable
+     * property and converts the property to an observable value.
+     */
+    fun <T> column(title: String, prop: KMutableProperty1<S, T>): TableColumn<S, T> {
+        val column = TableColumn<S, T>(title)
+        column.cellValueFactory = Callback { observable(it.value, prop) }
+        columns.add(column)
+        return column
+    }
+
+    /**
+     * Create a column with a value factory that extracts the value from the given property and
+     * converts the property to an observable value.
+     */
+    fun <T> column(title: String, prop: KProperty1<S, T>): TableColumn<S, T> {
+        val column = TableColumn<S, T>(title)
+        column.cellValueFactory = Callback { observable(it.value, prop) }
+        columns.add(column)
+        return column
+    }
+
+    /**
+     * Create a column with a value factory that extracts the observable value from the given function reference.
+     * This method requires that you have kotlin-reflect on your classpath.
+     */
+    inline fun <reified T> column(title: String, observableFn: KFunction<ObjectProperty<T>>): TableColumn<S, T> {
+        val column = TableColumn<S, T>(title)
+        column.cellValueFactory = ReflectionHelper.CellValueFunctionRefCallback(observableFn)
         columns.add(column)
         return column
     }