+++ src/main/java/tornadofx/Properties.kt
@@ -1,11 +1,10 @@
 package tornadofx
 
-import javafx.beans.property.ObjectProperty
-import javafx.beans.property.Property
-import javafx.beans.property.SimpleObjectProperty
+import javafx.beans.property.*
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.KMutableProperty1
 import kotlin.reflect.KProperty
+import kotlin.reflect.KProperty1
 
 fun <T> property(value: T? = null) = PropertyDelegate(SimpleObjectProperty<T>(value))
 fun <T> property(block: () -> Property<T>) = PropertyDelegate(block())
@@ -26,69 +25,26 @@ fun <T> Any.getProperty(prop: KMutableProperty1<*, T>): ObjectProperty<T> {
     // avoid kotlin-reflect dependency
     val field = this.javaClass.getDeclaredField("${prop.name}\$delegate")
     field.isAccessible = true
+    @Suppress("UNCHECKED_CAST")
     val delegate = field.get(this) as PropertyDelegate<T>
     return delegate.fxProperty as ObjectProperty<T>
 }
 
-enum class SingleAssignThreadSafetyMode {
-    SYNCHRONIZED,
-    NONE
+/**
+ * Convert an owner instance and a corresponding property reference into an observable
+ */
+fun <S, T> observable(owner: S, prop: KMutableProperty1<S, T>): ObjectProperty<T> {
+    return object : SimpleObjectProperty<T>(owner, prop.name) {
+        override fun get() = prop.get(owner)
+        override fun set(v: T) = prop.set(owner, v)
+    }
 }
 
-fun <T> singleAssign(threadSafeyMode: SingleAssignThreadSafetyMode = SingleAssignThreadSafetyMode.SYNCHRONIZED): SingleAssign<T> =
-    if (threadSafeyMode.equals(SingleAssignThreadSafetyMode.SYNCHRONIZED)) SynchronizedSingleAssign<T>() else UnsynchronizedSingleAssign<T>()
-
-private object UNINITIALIZED_VALUE
-
-interface SingleAssign<T> {
-    fun isInitialized(): Boolean
-    operator fun getValue(thisRef: Any?, property: KProperty<*>): T
-    operator fun setValue(thisRef: Any?, property: KProperty<*>, value: T)
-}
-
-private class SynchronizedSingleAssign<T>: SingleAssign<T> {
-
-    @Volatile
-    private var initialized = false
-
-    @Volatile
-    private var _value: Any? = UNINITIALIZED_VALUE
-
-    override operator fun getValue(thisRef: Any?, property: KProperty<*>): T {
-        if (!initialized)
-            throw Exception("Value has not been assigned yet!")
-        return _value as T
+/**
+ * Convert an owner instance and a corresponding property reference into an observable
+ */
+fun <S, T> observable(owner: S, prop: KProperty1<S, T>): ReadOnlyObjectProperty<T> {
+    return object : ReadOnlyObjectWrapper<T>(owner, prop.name) {
+        override fun get() = prop.get(owner)
     }
-
-    override operator fun setValue(thisRef: Any?, property: KProperty<*>, value: T) {
-        synchronized(this) {
-            if (initialized) {
-                throw Exception("Value has already been assigned!")
-            }
-            _value = value as T
-            initialized = true
-        }
-    }
-    override fun isInitialized() = initialized
-}
-
-private class UnsynchronizedSingleAssign<T>: SingleAssign<T> {
-
-    private var initialized = false
-    private var _value: Any? = UNINITIALIZED_VALUE
-
-    override operator fun getValue(thisRef: Any?, property: KProperty<*>): T {
-        if (!initialized)
-            throw Exception("Value has not been assigned yet!")
-        return _value as T
-    }
-
-    override operator fun setValue(thisRef: Any?, property: KProperty<*>, value: T) {
-        if (initialized) {
-            throw Exception("Value has already been assigned!")
-        }
-        _value = value as T
-        initialized = true
-    }
-    override fun isInitialized() = initialized
 }