+++ CHANGELOG.md
@@ -15,7 +15,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Builder for Tab now require the content pane as input parameter to avoid confusion (https://github.com/edvin/tornadofx/issues/8)
 - UIComponent#openModal() no longer requests focus for root component - caller can decide instead
-- Property delegate not produces non-nullable property types
+- Property delegate now produces non-nullable property types
 
 ## [1.3.0] - 2016-02-19