+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
-- `Node.gridpaneConstraints extension function (https://github.com/edvin/tornadofx/issues/12)
+- BorderPane builder (https://github.com/edvin/tornadofx/pull/16)
+- `Node.gridpaneConstraints` extension function (https://github.com/edvin/tornadofx/issues/12)
 - `Node.vboxConstraints` extension function
 - `Node.hboxConstraints` extension function
 - `TableView` builder (https://github.com/edvin/tornadofx/issues/11)
@@ -14,7 +15,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Builder for Tab now require the content pane as input parameter to avoid confusion (https://github.com/edvin/tornadofx/issues/8)
 - UIComponent#openModal() no longer requests focus for root component - caller can decide instead
-- Property delegate not produces non-nullable property types
+- Property delegate now produces non-nullable property types
 
 ## [1.3.0] - 2016-02-19