+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- Arbitrary properties per `Component` (https://github.com/edvin/tornadofx/issues/23)
 - `singleAssign()` property delegates (https://github.com/edvin/tornadofx/issues/17)
 - Tests for `singleAssign()`
 - BorderPane builder (https://github.com/edvin/tornadofx/pull/16)
@@ -18,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Builder for Tab now require the content pane as input parameter to avoid confusion (https://github.com/edvin/tornadofx/issues/8)
 - UIComponent#openModal() no longer requests focus for root component - caller can decide instead
 - Property delegate now produces non-nullable property types
+- `GridPane.row()` no long hogs the GridPane `userData` object to track rowId
 
 ## [1.3.0] - 2016-02-19